🚚 move: NavProfile shared -> Desktop
