+++ app/src/components/elements/NavBar/Desktop/NavBar.tsx
