+++ app/src/components/elements/NavBar/Desktop/NavProfile.tsx
